+++ src/deprecations.jl
@@ -1,1 +1,1 @@
-# @deprecate OSM.random_route! OSM.plan_random_route!
+# Some deprecations exist in submodules Pathfinding, OSM