+++ src/models/zombies.jl
@@ -27,7 +27,7 @@ function zombies(; seed = 1234)
         OSM.plan_random_route!(human, model; limit = 50) # try 50 times to find a random route
     end
     ## We'll add patient zero at a specific (latitude, longitude)
-    start = OSM.nearest_intersection((51.5328328, 9.9351811), model)
+    start = OSM.nearest_road((51.5328328, 9.9351811), model)
     finish = OSM.nearest_node((51.530876112711745, 9.945125635913511), model)
 
     speed = rand(model.rng) * 5.0 + 2.0 # Random speed from 2-7kmph
@@ -39,13 +39,13 @@ end
 
 function zombie_agent_step!(agent, model)
     ## Each agent will progress slightly along their route
-    move_along_route!(agent, model, agent.speed * model.dt)
+    distance_left = move_along_route!(agent, model, agent.speed * model.dt)
 
     if is_stationary(agent, model) && rand(model.rng) < 0.1
         ## When stationary, give the agent a 10% chance of going somewhere else
         OSM.plan_random_route!(agent, model; limit = 50)
         ## Start on new route
-        move_along_route!(agent, model, agent.speed * model.dt)
+        move_along_route!(agent, model, distance_left)
     end
 
     if agent.infected